fix the bug of delay time

+++ app/src/main/java/com/example/alan/customframe/latte/Configurator.java
@@ -75,7 +75,7 @@ public class Configurator {
         return this;
     }
 
-    public Configurator withLoaderTime(int time){
+    public Configurator withLoaderTime(long time){
         LATTE_CONFIGS.put(ConfigType.LOADER_DELAYED.name(),time);
         return this;
     }
